scrapers/crawl-scrape-pipeline/requirements.txt
# Core scraping frameworks
scrapy==2.11.0
scrapy-playwright==0.0.33
playwright>=1.40.0
# crawlee==0.1.0  # Commented out due to dependency conflict

# Data processing
pandas==2.1.4
pyyaml==6.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Environment & utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
pylint==3.0.3